requirements.txt
annotated-types==0.7.0
attrs==23.2.0
blinker==1.8.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.2
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==4.0.1
flask-openapi3==3.1.3
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.0.3
